+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## v0.0.3 - 2024-??-?? - ???
 
 * octodns-keyring tool added to enable secret management
-* KeyringSecrets.set method added, used by ^
+* KeyringSecrets.set & delete methods added, used by ^
 
 ## v0.0.2 - 2024-03-17 - What type is that thingy in the window